PM-1465 XPAY move async method to dedicated service class
